requirements.txt
jpype1==1.5.2
mysql-connector-python==9.3.0
packaging==25.0
pillow==11.2.1
pyreportjasper==2.1.4
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
shiboken6==6.9.0
